Assures the inputDelegate is not notified in the middle of dragging the cursor or extending a selection.
-moveCursorToPositionClosestToLocation:  changes to -moveCursorToPositionClosestToLocation:notifyInputDelegate:

Therefore, notifyInputDelegate can be NO in the methods for presenting, moving, and dismissing the Loupe.  Reusable methods are mashing together too many responsibilities.

The key is to wrap the long press and pan gesture recognizers with calls to the inputDelegate, selectionWillChange: when the recognizers begin and selectionDidChange: when the recognizers end.

Now autocorrect bubbles will dismiss appropriately if some form of dragging begins and the text system will not be calling expensive UITextInput methods while we are dragging the loupe around.