Added Hyperlink support and early Media browsing for iPad.
The media browsing will require a different row for iPhone. Pages has a separate insert module for media including tables, charts and shapes. My solution shouldn't be final. Also unable to hide back button from image picker controller since it doesn't immediately see it's in a popover. It's surprising this didn't crash with an assert that it should be on it's own popover on iPad.