A delegate method for finer control over pasted content.
(NSAttributedString *)editorView:willPasteText:inRange:

There was a need to have more control over what gets pasted into an editor view and to easily alter it.  While editorView:shouldChangeTextInRange:replacementText: allows a user to abort text changes by returning NO, the user must then alter the content and insert it.  It is also an overloaded method as it is called for all operations which change text.  The new method is called only for pasting.