Fix invalid web archive paste not inserting content into editor
If copying from Google Drive app and pasting into DTRichTextEditor, the paste board content detected is treated as an empty web archive. Instead of pasting nothing, this commit checks to see if the attributed string is valid. If so, it performs the current behavior. Otherwise it allows the paste method to insert the content as a string.